requirements.txt
asgiref==3.11.1
Django==4.2.29
django-cors-headers==4.9.0
djangorestframework==3.16.1
groq==0.4.2
gunicorn==20.1.0
psycopg2-binary==2.9.11
python-dotenv==0.19.2
setuptools==82.0.1
sqlparse==0.5.5
